add problem fixers

+++ cogs/censor.py
@@ -1,5 +1,6 @@
 import json
 
+import discord
 from discord.ext import commands
 from profanity_check import predict
 
@@ -20,6 +21,21 @@ class Censor(commands.Cog):
     @commands.Cog.listener('on_message')
     @commands.Cog.listener('on_message_edit')
     async def profanity_filter_ml(self, *args):
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if args[0].author.id in self.config["all_exempt_user_ids"]:
+            return
         message = args[-1]
         if self.config["enabled"] and self.config["profanity_filter_ml"] and predict([message.content]) == [1]:
             await message.delete()
@@ -29,6 +45,16 @@ class Censor(commands.Cog):
     @commands.Cog.listener('on_message')
     @commands.Cog.listener('on_message_edit')
     async def profanity_filter(self, *args):
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if args[0].author.id in self.config["all_exempt_user_ids"]:
+            return
         message = args[-1]
         if self.config["enabled"] and self.config["profanity_filter"]:
             for word in message.content.split():
@@ -40,6 +66,16 @@ class Censor(commands.Cog):
     @commands.Cog.listener('on_message')
     @commands.Cog.listener('on_message_edit')
     async def message_char_limit(self, *args):
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if args[0].author.id in self.config["all_exempt_user_ids"]:
+            return
         message = args[-1]
         if self.config["enabled"] and 0 < self.config["message_char_limit"] <= len(message.content):
             await message.delete()
@@ -49,6 +85,16 @@ class Censor(commands.Cog):
     @commands.Cog.listener('on_message')
     @commands.Cog.listener('on_message_edit')
     async def caps_limit(self, *args):
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if args[0].author.id in self.config["all_exempt_user_ids"]:
+            return
         message = args[-1]
         if self.config["enabled"] and self.config["caps_limit_enabled"]:
             cap_count = 0
@@ -63,6 +109,16 @@ class Censor(commands.Cog):
     @commands.Cog.listener('on_message')
     @commands.Cog.listener('on_message_edit')
     async def word_filter(self, *args):
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if args[0].author.id in self.config["all_exempt_user_ids"]:
+            return
         message = args[-1]
         if self.config["enabled"] and self.config["word_filter_enabled"] \
                 and message.channel.id not in self.config["word_filter_channel_exceptions_array_ids"]:
@@ -78,6 +134,16 @@ class Censor(commands.Cog):
     @commands.Cog.listener('on_message')
     @commands.Cog.listener('on_message_edit')
     async def domain_filter(self, *args):
+        for i in args[0].author.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+        if args[0].author.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        for i in args[0].author.roles:
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if args[0].author.id in self.config["all_exempt_user_ids"]:
+            return
         message = args[-1]
         if self.config["enabled"] and self.config["filter_domains"]:
             for word in message.content.split():
@@ -95,7 +161,17 @@ class Censor(commands.Cog):
                                   f'{word} {message.content.lower().count(word.lower())} time(s)')
 
     @commands.Cog.listener('on_member_update')
-    async def nick_censor(self, before, after):
+    async def nick_censor(self, before: discord.Member, after: discord.Member):
+        for i in after.roles:
+            if i.id in self.config["wordfilter_exception_role_ids"]:
+                return
+            if i.id in self.config["all_exempt_roles"]:
+                return
+        if after.id in self.config["wordfilter_exception_user_ids"]:
+            return
+        if after.id in self.config["all_exempt_user_ids"]:
+            return
+
         if self.config["enabled"] and self.config["nickname_filter_enabled"]:
             for word in after.nick.lower().split():
                 if word.lower() in self.blocked_words:
